+++ inlang/source-code/versioned-interfaces/project-settings/src/interface.ts
@@ -25,7 +25,7 @@ export const _MessageLintRuleLevel = Type.Union([Type.Literal("error"), Type.Lit
  * ---------------- Settings ----------------
  */
 
-const InternalSettings = Type.Object({
+const InternalProjectSettings = Type.Object({
 	$schema: Type.Optional(Type.Literal("https://inlang.com/schema/project-settings")),
 	sourceLanguageTag: LanguageTag,
 	languageTags: Type.Array(LanguageTag, { uniqueItems: true }),
@@ -63,7 +63,7 @@ const InternalSettings = Type.Object({
 				"https://cdn.jsdelivr.net/npm/@inlang/plugin-csv@1/dist/index.js",
 				"./local-testing-plugin.js",
 			],
-		},
+		}
 	),
 	messageLintRuleLevels: Type.Optional(
 		Type.Record(_MessageLintRuleId, _MessageLintRuleLevel, {
@@ -74,18 +74,19 @@ const InternalSettings = Type.Object({
 					"messageLintRule.inlang.patternInvalid": "warning",
 				},
 			],
-		}),
+		})
 	),
 })
 
 /**
  * Settings defined via apps, plugins, lint rules, etc.
  */
-const ExternalSettings = Type.Record(
+export type ExternalProjectSettings = Static<typeof ExternalProjectSettings>
+export const ExternalProjectSettings = Type.Record(
 	Type.String({
 		// pattern includes ProjectSettings keys
 		pattern: `^((messageLintRule|plugin|app|library)\\.([a-z][a-zA-Z0-9]*)\\.([a-z][a-zA-Z0-9]*(?:[A-Z][a-z0-9]*)*)|\\$schema|${Object.keys(
-			InternalSettings.properties,
+			InternalProjectSettings.properties
 		)
 			.map((key) => key.replaceAll(".", "\\."))
 			.join("|")})$`,
@@ -99,8 +100,8 @@ const ExternalSettings = Type.Record(
 	// intersection between `InternalSettings`, which contains an array,
 	// and `ExternalSettings` which are objects possible
 	JSON as unknown as typeof JSONObject,
-	{ additionalProperties: false, description: "Settings defined by apps, plugins, etc." },
+	{ additionalProperties: false, description: "Settings defined by apps, plugins, etc." }
 )
 
 export type ProjectSettings = Static<typeof ProjectSettings>
-export const ProjectSettings = Type.Intersect([InternalSettings, ExternalSettings])
+export const ProjectSettings = Type.Intersect([InternalProjectSettings, ExternalProjectSettings])