+++ inlang/source-code/plugins/json/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @inlang/plugin-json
 
+## 4.6.0
+
+### Minor Changes
+
+- f40ab4ca9: refactor: use plugin api v2
+
+### Patch Changes
+
+- Updated dependencies [0f9dc72b3]
+  - @inlang/sdk@0.9.0
+
+## 4.5.0
+
+### Minor Changes
+
+- b7dfc781e: change message format match from object to array
+
+### Patch Changes
+
+- Updated dependencies [b7dfc781e]
+  - @inlang/sdk@0.8.0
+
 ## 4.4.0
 
 ### Minor Changes